+++ src/app/locales/en/components/archive-browser/table-head-row.js
@@ -1,4 +1,4 @@
-import headers from '../fb-columns-name/archive';
+import headers from '../-browser-columns-name/archive';
 
 export default {
   headers,