+++ src/app/locales/en/components/columns-configuration-popover/column-item.js
@@ -1,9 +1,9 @@
-import archiveFilesystem from '../fb-columns-name/archive-filesystem';
-import archive from '../fb-columns-name/archive';
-import dataset from '../fb-columns-name/dataset';
-import filesystem from '../fb-columns-name/filesystem';
-import shareFilesystem from '../fb-columns-name/share-filesystem';
-import transfer from '../fb-columns-name/transfer';
+import archiveFilesystem from '../-browser-columns-name/archive-filesystem';
+import archive from '../-browser-columns-name/archive';
+import dataset from '../-browser-columns-name/dataset';
+import filesystem from '../-browser-columns-name/filesystem';
+import shareFilesystem from '../-browser-columns-name/share-filesystem';
+import transfer from '../-browser-columns-name/transfer';
 
 export default {
   noViewTip: 'Column hidden due to limited horizontal space. Try resizing the browser window, decreasing zoom or disabling other columns.',