+++ src/app/locales/en/components/filesystem-browser/table-head-row.js
@@ -1,4 +1,4 @@
-import headers from '../fb-columns-name/filesystem';
+import headers from '../-browser-columns-name/filesystem';
 
 export default {
   headers,